Mapo Tofu is now ready
git-svn-id: http://svn.code.sf.net/p/iso2mesh/code/trunk/iso2mesh@282 786e58fb-9377-0410-9ff7-e4ac0ac0635c

+++ vol2surf.m
@@ -84,9 +84,9 @@ if(~isempty(img))
 
     for i=1:maxlevel
       if(i<maxlevel)
-          levelmask=(newimg>=isovalues(i) & newimg<isovalues(i+1));
+          levelmask=int8(newimg>=isovalues(i) & newimg<isovalues(i+1));
       else
-          levelmask=(newimg>=isovalues(i));
+          levelmask=int8(newimg>=isovalues(i));
       end
       [levelno,levelel]=binsurface(levelmask);
       if(~isempty(levelel))
